requirements.txt
mysql-connector-python~=8.3.0
fastapi~=0.112.0
passlib~=1.7.4

python-jose~=3.3.0
uvicorn~=0.29.0